Prioritize the present working control repository
Without this patch the control repository in the present working directory is
not used unless explicitly added to the personal configuration file.  This is a
problem because the personal configuration file should be the same across all
teams, and the infra team will be switching among teams quite a bit.

This patch addresses the problem by implementing the following behaviors.  If
the pwd is inside a control repository (denoted by the presence of
.rizzo.json), that control repository is added, or moved, to the top of the
control repo list.

The personal configuration at ~/.rizzo.json is not considered a project control
repository.