WIP - Refactor config handling
Adding tests for the behavior of moving the current working directory control
repo to the top of the list was more difficult than expected.  Mocking out the
configuration file reading proved to be difficult, so this patch aims to
consolidate configuration handling at the application layer rather than as a
mix-in to each subcommand.

This work is being abandoned because it's taking too long.  Only two tests need
to be added to move forward, so saving this work as in-progress.

Pick back up working on the individual run methods for each subcommand and
sorting out how to inject the configuration.